Create directory for -alpine and -slim Dockerfile
To support multiple architectures of offical images build on Docker
Hub, we need to use RFC 2822 based format for manifest file.

 https://github.com/docker-library/official-images#instruction-format

In this format, if Directory is present, Bashbrew will look for the
Dockerfile inside the specified subdirectory.  And there is no way to
find a Dockerfile with customized name.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>